+++ core/src/gaia/cu9/ari/gaiaorbit/scenegraph/Star.java
@@ -69,11 +69,11 @@ public class Star extends Particle {
         this.parentName = ROOT_NAME;
     }
 
-    public Star(Vector3d pos, float appmag, float absmag, float colorbv, String name, long starid) {
+    public Star(Vector3d pos, float appmag, float absmag, float colorbv, String name, int starid) {
         super(pos, appmag, absmag, colorbv, name, starid);
     }
 
-    public Star(Vector3d pos, float appmag, float absmag, float colorbv, String name, double ra, double dec, long starid) {
+    public Star(Vector3d pos, float appmag, float absmag, float colorbv, String name, float ra, float dec, int starid) {
         super(pos, appmag, absmag, colorbv, name, ra, dec, starid);
     }
 
@@ -97,17 +97,6 @@ public class Star extends Particle {
                         camera.checkClosest(this);
                         addToRender(this, RenderGroup.MODEL_S);
                     }
-                    // Check Sol position for gravity distortion
-
-                    //                    if (this.name.equalsIgnoreCase("sol")) {
-                    //                        // We have the closest shader star
-                    //                        Vector3 aux = auxVector3f.get();
-                    //                        camera.getCamera().project(aux.set(transform.getTranslationf()));
-                    //                        float x = aux.x;
-                    //                        float y = aux.y;
-                    //
-                    //                        EventManager.instance.post(Events.GRAVITATIONAL_LENSING_PARAMS, x, y);
-                    //                    }
                 }
             }
             if (renderText()) {
